+++ packages/plugins/Attributions/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 
 - Feature: Use the configured `layoutTag` to identify the direction the attributions content should open in `renderType` `independent`.
+- Feature: Add configuration parameter `buttonComponent` to use a different component as the button to open the attributions if `renderType` is set to `'independent'`.
 
 ## 1.4.0