+++ packages/types/custom/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Feature: Add new interface `LayerChooserConfiguration` and the parameter `layerChooser` to `MapConfig` for the configurability of `@polar/plugin-layer-chooser`.
 - Feature: Add new parameter `component` to `IconMenuConfiguration` for the configurability of the shown component.
 - Feature: Add new parameter `component` to `ZoomConfiguration` for the configurability of the shown component.
+- Feature: Add new parameter `buttonComponent` to `AttributionsConfiguration` for the configurability of the button component in `renderType` `independent`.
 
 ## 2.0.0